4.15.-4.21. done

+++ backend/controllers/blogs.js
@@ -1,26 +1,47 @@
-const notesRouter = require('express').Router()
+const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
-notesRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
+
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', { username:1, name:1 })
   response.json(blogs)
 })
 
-notesRouter.post('/',  async (request, response, next) => {
+blogsRouter.post('/',  async (request, response, next) => {
     const body = request.body
+    const token = request.token
+
+    
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
   
     const blog = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: 0
+      user: user._id
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
 })
 
-notesRouter.get('/:id', async (request, response, next) => {
+blogsRouter.get('/:id', async (request, response, next) => {
     const blog = await Blog.findById(request.params.id)
     if (blog) {
       response.json(blog)
@@ -29,12 +50,23 @@ notesRouter.get('/:id', async (request, response, next) => {
     }
 })
 
-notesRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if ( blog.user.toString() === user._id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  
 })
 
-notesRouter.put('/:id', async (request, response, next) => {
+blogsRouter.put('/:id', async (request, response, next) => {
   const blog = {
     title: request.body.title,
     author: request.body.author,
@@ -44,4 +76,4 @@ notesRouter.put('/:id', async (request, response, next) => {
   await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
   response.status(201).end()
 })
-module.exports = notesRouter
+module.exports = blogsRouter
